Implement more methods in RequestContextWrapper.
