+++ core/src/main/java/com/linecorp/armeria/client/ClientRequestContextWrapper.java
@@ -136,27 +136,6 @@ public class ClientRequestContextWrapper
     }
 
     @Override
-    public void cancel(Throwable cause) {
-        delegate().cancel(cause);
-    }
-
-    @Override
-    public void cancel() {
-        delegate().cancel();
-    }
-
-    @Override
-    public void timeoutNow() {
-        delegate().timeoutNow();
-    }
-
-    @Nullable
-    @Override
-    public Throwable cancellationCause() {
-        return delegate().cancellationCause();
-    }
-
-    @Override
     public boolean isCancelled() {
         return delegate().isCancelled();
     }
