+++ core/src/main/java/com/linecorp/armeria/server/ServiceRequestContextWrapper.java
@@ -156,27 +156,6 @@ public class ServiceRequestContextWrapper
     }
 
     @Override
-    public void cancel(Throwable cause) {
-        delegate().cancel(cause);
-    }
-
-    @Override
-    public void cancel() {
-        delegate().cancel();
-    }
-
-    @Override
-    public void timeoutNow() {
-        delegate().timeoutNow();
-    }
-
-    @Nullable
-    @Override
-    public Throwable cancellationCause() {
-        return delegate().cancellationCause();
-    }
-
-    @Override
     public boolean isCancelled() {
         return delegate().isCancelled();
     }
